# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OtpRequest {
  /** Email address to send the OTP to */
  email: string;
}

export interface OtpSendResult {
  success: boolean;
  message: string;
  /**
     * OTP shown only in development mode
     * @nullable
     */
  devOtp?: string | null;
}

export interface OtpVerifyRequest {
  /** Email address the OTP was sent to */
  email: string;
  /** 4-digit OTP */
  otp: string;
}

export interface OtpVerifyResult {
  success: boolean;
  message: string;
  /** Short-lived opaque token. Pass as "Authorization: Bearer <token>" on all profile endpoints. Valid for 30 minutes. */
  verificationToken: string;
}

export interface WelcomeRequest {
  /** Verified email address of the new user */
  email: string;
  /** Display name chosen during profile setup */
  name: string;
  /** Optional bio text entered during profile setup */
  about?: string;
}

export interface WelcomeResult {
  success: boolean;
  message: string;
}

export interface ErrorResult {
  error: string;
}

export interface UpsertProfileRequest {
  /** Display name */
  name: string;
  /** Short bio / about text */
  about?: string;
  /** Profile photo as a data URL */
  photo?: string;
}

export interface ProfileResult {
  success: boolean;
  message: string;
}

export interface UpdateLocationRequest {
  /** WGS-84 latitude in decimal degrees */
  latitude: number;
  /** WGS-84 longitude in decimal degrees */
  longitude: number;
}

export interface LocationResult {
  success: boolean;
  message: string;
}

export interface NearbyProfileCard {
  name: string;
  /** Profile photo data URL */
  photo: string;
  /** Great-circle distance from the requester in metres */
  distanceMeters: number;
  /** Short AI-generated conversation-starter line based on the person's bio */
  conversationStarter: string;
}

export interface NearbyProfilesResult {
  profiles: NearbyProfileCard[];
}