# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to email address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpSendResult"
        "400":
          description: Invalid email address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
  /auth/send-welcome:
    post:
      operationId: sendWelcome
      tags: [auth]
      summary: Send onboarding welcome email after profile completion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WelcomeRequest"
      responses:
        "200":
          description: Welcome email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WelcomeResult"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify 4-digit OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyRequest"
      responses:
        "200":
          description: OTP verified — response includes a short-lived verification token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResult"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
...
components:
  securitySchemes:
    verificationToken:
      type: http
      scheme: bearer
      description: >
        Short-lived opaque token issued by POST /auth/verify-otp.
        Valid for 30 minutes. Pass as "Authorization: Bearer <token>".
  schemas:
...
    OtpRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          description: Email address to send the OTP to
    OtpSendResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        devOtp:
          type: ["string", "null"]
          description: OTP shown only in development mode
    OtpVerifyRequest:
      type: object
      required: [email, otp]
      properties:
        email:
          type: string
          description: Email address the OTP was sent to
        otp:
          type: string
          description: 4-digit OTP
    OtpVerifyResult:
      type: object
      required: [success, message, verificationToken]
      properties:
        success:
          type: boolean
        message:
          type: string
        verificationToken:
          type: string
          description: >
            Short-lived opaque token. Pass as "Authorization: Bearer <token>"
            on all profile endpoints. Valid for 30 minutes.
    WelcomeRequest:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
          description: Verified email address of the new user
        name:
          type: string
          description: Display name chosen during profile setup
        about:
          type: string
          description: Optional bio text entered during profile setup
    WelcomeResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResult:
      type: object
      required: [error]
      properties:
        error:
          type: string
...