# lib/api-zod/src/generated/types/welcomeRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WelcomeRequest {
  /** Verified email address of the new user */
  email: string;
  /** Display name chosen during profile setup */
  name: string;
  /** Optional bio text entered during profile setup */
  about?: string;
}